Fix code scanning alert no. 1: Comparison of narrow type with wide type in loop condition

+++ src/tcs/ps/ps_utils.c
@@ -345,7 +345,7 @@ int
 init_disk_cache(int fd)
 {
 	UINT32 num_keys = get_num_keys_in_file(fd);
-	UINT16 i;
+	UINT32 i;
 	UINT64 tmp_offset;
 	int rc = 0, offset;
 	struct key_disk_cache *tmp, *prev = NULL;
